improve ccsds sp test structure
